+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+TBD
+===
+
+Features
+--------
+
+* [#114](https://github.com/dotnetinstaller/dotnetinstaller/issues/114) - Added support for operating system product type filters - [@icnocop](https://github.com/icnocop).
+
 3.2.115 (1/15/2022)
 ===================